Extract goalsComponent to prevent circular dependency
